refactor : seat, seatScheduleInfo 파일 패키지 위치 변경
