+++ src/main/java/org/example/siljeun/domain/seatscheduleinfo/dto/request/SeatScheduleUpdateInfoRequest.java
@@ -1,4 +1,4 @@
-package org.example.siljeun.domain.seat.dto.request;
+package org.example.siljeun.domain.seatscheduleinfo.dto.request;
 
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
