+++ src/main/java/org/example/siljeun/domain/seatscheduleinfo/scheduler/SeatStatusPreloaderScheduler.java
@@ -1,10 +1,10 @@
-package org.example.siljeun.domain.schedule.scheduler;
+package org.example.siljeun.domain.seatscheduleinfo.scheduler;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.example.siljeun.domain.schedule.entity.Schedule;
 import org.example.siljeun.domain.schedule.repository.ScheduleRepository;
-import org.example.siljeun.domain.schedule.repository.SeatScheduleInfoRepository;
+import org.example.siljeun.domain.seat.repository.SeatScheduleInfoRepository;
 import org.example.siljeun.domain.seat.entity.SeatScheduleInfo;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.scheduling.annotation.Scheduled;
